fix prototype parameter type

+++ dnd/src/linkst.h
@@ -18,7 +18,7 @@
 int linkst_joined(int idx_a, int idx_b, uint8_t **adjacency_matrix, int max_node);
 int linkst_join(int idx_a, int idx_b, uint8_t **adjacency_matrix, int max_node);
 void linkst_disjoin(int idx, uint8_t **adjacency_matrix, int active_nodes);
-void linkst_free(uint8_t **adjacency_matrix, size_t max_node);
-uint8_t **linkst_new(size_t max_node);
+void linkst_free(uint8_t **adjacency_matrix, uint32_t max_node);
+uint8_t **linkst_new(uint32_t max_node);
 
 #endif
